Assignment-1/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-64
appnope=0.1.0=py36_1000
attrs=18.2.0=py_0
backcall=0.1.0=py_0
biopython=1.72=py36h470a237_0
blas=1.0=mkl
bleach=3.0.2=py_0
bzip2=1.0.6=1
ca-certificates=2018.10.15=ha4d7672_0
certifi=2018.10.15=py36_1000
decorator=4.3.0=py_0
entrypoints=0.2.3=py36_1002
icu=58.2=hfc679d8_0
intel-openmp=2019.0=118
ipykernel=5.1.0=py36h24bf2e0_1001
ipython=7.1.1=py36h24bf2e0_1000
ipython_genutils=0.2.0=py_1
ipywidgets=7.4.2=py_0
jedi=0.13.1=py36_1000
jinja2=2.10=py_1
jpeg=9c=h470a237_1
jsonschema=3.0.0a3=py36_1000
jupyter=1.0.0=py_1
jupyter_client=5.2.3=py_1
jupyter_console=6.0.0=py_0
jupyter_core=4.4.0=py_0
libffi=3.2.1=hfc679d8_5
libgfortran=3.0.1=h93005f0_2
libpng=1.6.35=ha92aebf_2
libsodium=1.0.16=h470a237_1
markupsafe=1.1.0=py36h470a237_0
mistune=0.8.4=py36h470a237_0
mkl=2019.0=118
mkl_fft=1.0.6=py36_0
mkl_random=1.0.2=py36_0
nbconvert=5.3.1=py_1
nbformat=4.4.0=py_1
ncurses=6.1=hfc679d8_1
notebook=5.7.0=py36_1000
numpy=1.15.4=py36h6a91979_0
numpy-base=1.15.4=py36h8a80b8c_0
openssl=1.0.2p=h470a237_1
pandoc=2.3.1=0
pandocfilters=1.4.2=py_1
parso=0.3.1=py_0
pexpect=4.6.0=py36_1000
pickleshare=0.7.5=py36_1000
pip=18.1=py36_1000
prometheus_client=0.4.2=py_0
prompt_toolkit=2.0.7=py_0
ptyprocess=0.6.0=py36_1000
pygments=2.2.0=py_1
pyqt=5.6.0=py36h8210e8a_7
pyrsistent=0.14.5=py36h470a237_1
python=3.6.6=h5001a0f_0
python-dateutil=2.7.5=py_0
pyzmq=17.1.2=py36hae99301_1
qt=5.6.2=h4e759b2_9
qtconsole=4.4.3=py_0
readline=7.0=haf1bffa_1
send2trash=1.5.0=py_0
setuptools=40.6.2=py36_0
sip=4.18.1=py36hfc679d8_0
six=1.11.0=py36_1001
sqlite=3.25.3=hb1c47c0_0
terminado=0.8.1=py36_1001
testpath=0.4.2=py36_1000
tk=8.6.8=ha92aebf_0
tornado=5.1.1=py36h470a237_0
traitlets=4.3.2=py36_1000
wcwidth=0.1.7=py_1
webencodings=0.5.1=py_1
wheel=0.32.2=py36_0
widgetsnbextension=3.4.2=py36_1000
xz=5.2.4=h470a237_1
zeromq=4.2.5=hfc679d8_5
zlib=1.2.11=h470a237_3
